+++ compiler/src/main/scala/edg/compiler/CompilerError.scala
@@ -1,0 +1,77 @@
+package edg.compiler
+
+import edg.EdgirUtils.SimpleLibraryPath
+import edg.wir.{DesignPath, IndirectDesignPath}
+import edgir.expr.expr
+import edgir.ref.ref
+
+
+class ElaboratingException(msg: String, wrapped: Exception) extends Exception(f"$msg:\n$wrapped")
+
+
+sealed trait CompilerError
+
+object CompilerError {
+  case class Unelaborated(elaborateRecord: ElaborateRecord, missing: Set[ElaborateRecord]) extends CompilerError {
+    // These errors may be redundant with below, but provides dependency data
+    override def toString: String = s"Unelaborated missing dependencies $elaborateRecord:\n" +
+        s"${missing.map(x => s"- $x").mkString("\n")}"
+  }
+  case class LibraryElement(path: DesignPath, target: ref.LibraryPath) extends CompilerError {
+    override def toString: String = s"Unelaborated library element ${target.toSimpleString} @ $path"
+  }
+  case class UndefinedPortArray(path: DesignPath, portType: ref.LibraryPath) extends CompilerError {
+    override def toString: String = s"Undefined port array ${portType.toSimpleString} @ $path"
+  }
+
+  case class LibraryError(path: DesignPath, target: ref.LibraryPath, err: String) extends CompilerError {
+    override def toString: String = s"Library error ${target.toSimpleString} @ $path: $err"
+  }
+  case class GeneratorError(path: DesignPath, target: ref.LibraryPath, err: String) extends CompilerError {
+    override def toString: String = s"Generator error ${target.toSimpleString} @ $path: $err"
+  }
+  case class RefinementSubclassError(path: DesignPath, refinedLibrary: ref.LibraryPath, designLibrary: ref.LibraryPath)
+      extends CompilerError {
+    override def toString: String =
+      s"Invalid refinement ${refinedLibrary.toSimpleString} <- ${designLibrary.toSimpleString} @ $path"
+  }
+
+  case class OverAssign(target: IndirectDesignPath,
+                        causes: Seq[OverAssignCause]) extends CompilerError {
+    override def toString: String = s"Overassign to $target:\n" +
+        s"${causes.map(x => s"- $x").mkString("\n")}"
+  }
+
+  case class BadRef(path: DesignPath, ref: IndirectDesignPath) extends CompilerError
+
+  case class AbstractBlock(path: DesignPath, blockType: ref.LibraryPath) extends CompilerError {
+    override def toString: String = s"Abstract block: $path (of type ${blockType.toSimpleString})"
+  }
+
+  case class FailedAssertion(root: DesignPath, constrName: String,
+                             value: expr.ValueExpr, result: ExprValue) extends CompilerError {
+    override def toString: String =
+      s"Failed assertion: $root.$constrName, ${ExprToString.apply(value)} => $result"
+  }
+  case class MissingAssertion(root: DesignPath, constrName: String,
+                              value: expr.ValueExpr, missing: Set[IndirectDesignPath]) extends CompilerError {
+    override def toString: String =
+      s"Unevaluated assertion: $root.$constrName (${ExprToString.apply(value)}), missing ${missing.mkString(", ")}"
+  }
+
+  // TODO should this be an error? Currently a debugging tool
+  case class EmptyRange(param: IndirectDesignPath, root: DesignPath, constrName: String,
+                        value: expr.ValueExpr) extends CompilerError
+
+  sealed trait OverAssignCause
+  object OverAssignCause {
+    case class Assign(target: IndirectDesignPath, root: DesignPath, constrName: String, value: expr.ValueExpr)
+        extends OverAssignCause {
+      override def toString = s"Assign $target <- ${ExprToString(value)} @ $root.$constrName"
+    }
+    case class Equal(target: IndirectDesignPath, source: IndirectDesignPath)  // TODO constraint info once we track that?
+        extends OverAssignCause {
+      override def toString = s"Equals $target = $source"
+    }
+  }
+}
